requirements.txt
openpyxl==3.1.2
python-docx==1.1.0
python-dotenv==1.0.1
requests==2.32.3
